move metadata out of CreateOrUpdate
Since we're handling this separatly with a specific store method, we
don't want to add a second place to set metadata here.

If/when we update things using CreateOrUpdate, we'll make a Tx method
for it. This way we don't need to update call sites across the app for
the new signature.