update SetMetadata to support service permissions validation
So this query is a bit more complicated, but we want to make sure no
future bug allows a service/integration key to update the metadata of an
alert that doesn't belong to it.

Integration keys in GoAlert are tied to a single service, and all DB
calls regarding alerts should be guarded such that only alerts belonging
to that specific service may be modified, including the newly introduced
metadata.

The approach I took here allows us to pass a service ID, optionally
(i.e., ONLY when a request comes from an integration), where the query
will fail to insert/update any data if it's mismatched. If the service
ID is null, it will skip this check (i.e., a user action).