move store methods to metadata.go
For organization sake, moving the store methods to a separate file (we
have a bad habbit of having the store.go files growing forever).

I also implemented things using the `gadb` and taking a `DBTX` rather
than a sql.Tx. This is the most recent pattern, although most DB code
has not been refactored for this.

I'm using the `map[string]string` Go type directly to get rid of any
indirection on the metadata format external to this package. Internally,
I added a `metadataDBFormat` that hopefully by name makes it clearer
it's only intended to be used at rest.

Compared to what was there before, I also:
- Added assertions on the type string
- Added a MetadataAlertID type that includes an alert ID for the
  dataloader
- Added an Access Denied error if SetMetadataTx fails because of an
  invalid alert or service ID mismatch