move validation to metadata.go
For this I just moved the validation to metadata.go and implemented it
as a standalone function rather than a method.

This was a mistake on my part of recommending a method -- we really
don't need a specific type for metadata as it's just a
`map[string]string` we also don't have a use-case to "normalize" the
data, as we always store it as-is.

This is different to things like services, schedules, etc... where we
will do things like trim whitespace, remove double-spaces, to
"normalize" the input.

For this it's just pass or fail.