move map handling out of Tx function
This is a bit of an optimization -- we can handle allocating the memory, and processing the map outside of the transaction to keep the transaction as quick as possible.

Additionally, we can do some early validation to jump out _before_ we get to the withContextTx call.

This helps because to create the alert, we actually:
- Start a transaction
- Grab a global exclusive lock on the service's row
- Create the alert itself
- Add the log entry about the alert

So if we already know the metadata is going to be rejected, we can skip that.

Lastly, now that the alert store only takes `map[string]map` it simplifies the code here as well.