fixed daily rotation participation span for schedules with day restrictions
When a daily rotation is combined with a shedule which has
restriction(eg: active only on weekdays) the rotation advances through
calendar time and not with active days. Due to this some participants
fell on the inactivate days and never appeared on the schedule.

This PR fixes it by advancing rotations based on active day instead of
calendar time.

Fixes: https://github.com/target/goalert/issues/4447